requirements.txt
mysqlclient
flask==2.0.1
Flask-Login==0.6.2
flask_migrate==3.1.0
flask_wtf==0.15.1
flask_sqlalchemy==2.5.1
sqlalchemy==1.4.23
flask_bcrypt==0.7.1
jinja2==3.0.1
python-decouple==3.4
email_validator==1.1.3
Werkzeug==2.0.3

